Working on the only human project.
Working on the only human project.
Solved the problem of the coupon with the single product buy.
Display the message of discounted price into the product detail page.
Done the functionality of the coupon.
Now, user can only use the coupon for single time.
If, user get wants used one coupon for multiple time then, it shows the
error msg.
Work on the edit, add functionality of the product.
Done the functionality add, edit functionality with additional charges
of product into product management.
Work on the prices type(Additional/less).
Saved the data of additional/less prices into the database with product
reference.
Shown the price of the product with additional charges.
Work on the functionality of the 'unavailable product'. if product is
not available then, it sends the email to user.
Working on the notification of the unavailable product.